chore: add issue status checker and script checking
We often link to opened issues in the code or the documentation.
But we sometimes forget to remove the comment once the issue is closed.
This commit adds a simple checker to check the status of all referenced
issues and adds it to `make pcc`.

Also it appeared that we didn't check our scripts for formatting.
This is now done.