chore(ci): pass release tag and build script to slab
Doing this allows handling non-linearity of release tag version.
For example if a v0.6.2 is released after a 0.7.0 (latest release
in Pypi), then an AWS AMI will be created with concrete-ml v0.6.2.
Prior to this commit, the AMI created would contained the latest
concrete-ml version available in Pypi, thus 0.7.0 in the previous
example.