chore: update onnxoptimizer version
Our current version of onnxoptimizer is not up to date.
The version that we currently have does not have a build wheel for ARM
chips. The newer version of onnxoptimizer does have a built wheel.
This commit updated the version of onnxoptimizer so that builds on ARM
based system can be faster.