chore: handle version not available in deployment
We currently are in a situation where when updating the version number
of Concrete-ML the CI will break because the deployment test will try to
install the new version while it's not available anywhere.

To tackle this we can relax the version constraint if pip can't find the
newer version.