fix: fix flaky padding test
The issue was that we assumed that none of the values in the test
input were mapped to 0 by the quantization.
This assumption was proved to be false.
A simple way to force that no value is set to 0 by the quantizer, learned
over [0,1], is to simply set the test input to 1 everywhere.