chore: change the way we use gitleaks
Until now we used gitleaks to check the whole history but it makes more
sense to only check from the first commit to the current commit.

Otherwise it would break all PRs if someone pushed something that was
detected as a leak.
This introduces some issues also when working with multiple remotes.

The commit changes the behavior to check from first commit to current
commit.