fix: fix concurrency issue in release process
In the current release process the concurrency group for release actions
is just the name of the github action.
That means that we can't do a patch release and a minor release at the
same, which is an issue.

I  propose to solve this by adding the github.ref to the concurrency
group.

https://docs.github.com/en/actions/learn-github-actions/contexts#github-context