chore: still experimenting on TLU adjustment
Still experimenting with it. In the current setting using our approach
instead of the 6-bit rounding everywhere approach we degrade the top-1
accuracy of the CIFAR model too much (from 0.88 with the torch model, to
0.86 with rounding 6-bit to 0.79 for auto-rounding).

I added a notebook to visualize what happens with the auto-rounding
optimization on a simple setting. Things look fishy to me.

Let's keep iterating on it until we find an appropriate setting that
works on the CIFAR-10 use-case and at least another use-case.