wip: all tests passed once
All tests passed once.

Main improvement was a change on the closed form used for the bias.
There is probably still something to improve there.
I also added clipping after down-scaling in the LUT and a check on delta
to see if something could be gained. My guess is that there is still
something to gain in this specific situtation if the
ceil(log2(x_max-x_min)) < bit_width([x_max, x_min]).
In this specific situation we could indeed just center everything and
scale to match this observed bit-width.