Some suggestions on updating text: Update linear.md
Hey, I have some suggestions on this text:

There is a typo in "alllows"
There are three 'l's here.
Corrected: "allows"

"similar the one of"
This is awkward phrasing. It should be:
Corrected: "similar to the one of"

"helping us enhance our documentation"
This could be clearer as "help us enhance."
Corrected: "help us enhance our documentation"

Thanks.